requirements.txt
Django==1.9
Pillow==3.0.0
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.22
coverage==4.0.3
django-appconf==1.0.1
django-celery==3.1.17
django-debug-toolbar==1.4
django-extensions==1.6.1
django-imagekit==3.3
django-pipeline==1.5.4
django-redis-sessions==0.5.0
django-registration-redux==1.3
django-super-inlines==0.1.4
django-webpack-loader==0.2.4
django-websocket-redis==0.4.6
djangorestframework==3.3.2
djoser==0.4.3
drf-nested-routers==0.11.1
funcsigs==0.4
futures==3.0.3
gevent==1.0.2
greenlet==0.4.9
kombu==3.0.34
mercurial==3.1.1
mock==1.3.0
pbr==1.8.1
pilkit==1.1.13
psycopg2==2.6.1
py==1.4.31
pytest==2.8.5
pytest-cov==2.2.1
pytest-django==2.9.1
python-dateutil==2.5.0
pytz==2015.7
redis==2.10.5
requests==2.9.1
requests-mock==0.7.0
six==1.10.0
wsgiref==0.1.2
